third_party/blink/web_tests/platform/generic/fast/js/date-proto-generic-invocation-expected.txt

This test verifies that the functions of the Date prototype object are not generic, as documented in ECMA-262 rev3 section 15.9.5 Properties of the Date Prototype Object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS o.__proto__.toString() threw exception TypeError: Method Date.prototype.toString called on incompatible receiver [object Object].
PASS o.__proto__.toDateString() threw exception TypeError: Method Date.prototype.toDateString called on incompatible receiver [object Object].
PASS o.__proto__.toTimeString() threw exception TypeError: Method Date.prototype.toTimeString called on incompatible receiver [object Object].
PASS o.__proto__.toGMTString() threw exception TypeError: Method Date.prototype.toUTCString called on incompatible receiver [object Object].
PASS o.__proto__.toUTCString() threw exception TypeError: Method Date.prototype.toUTCString called on incompatible receiver [object Object].
PASS o.__proto__.toLocaleString() threw exception TypeError: Method Date.prototype.toLocaleString called on incompatible receiver [object Object].
PASS o.__proto__.toLocaleDateString() threw exception TypeError: Method Date.prototype.toLocaleDateString called on incompatible receiver [object Object].
PASS o.__proto__.toLocaleTimeString() threw exception TypeError: Method Date.prototype.toLocaleTimeString called on incompatible receiver [object Object].
PASS o.__proto__.getTime() threw exception TypeError: this is not a Date object..
PASS o.__proto__.getYear() threw exception TypeError: Method Date.prototype.getYear called on incompatible receiver [object Object].
PASS o.__proto__.getFullYear() threw exception TypeError: this is not a Date object..
PASS o.__proto__.getMonth() threw exception TypeError: this is not a Date object..
PASS o.__proto__.getDate() threw exception TypeError: this is not a Date object..
PASS o.__proto__.getDay() threw exception TypeError: this is not a Date object..
PASS o.__proto__.getHours() threw exception TypeError: this is not a Date object..
PASS o.__proto__.getMinutes() threw exception TypeError: this is not a Date object..
PASS o.__proto__.getSeconds() threw exception TypeError: this is not a Date object..
PASS o.__proto__.getMilliseconds() threw exception TypeError: this is not a Date object..
PASS o.__proto__.getTimezoneOffset() threw exception TypeError: this is not a Date object..
PASS o.__proto__.setTime() threw exception TypeError: Method Date.prototype.setTime called on incompatible receiver [object Object].
PASS o.__proto__.setMilliseconds() threw exception TypeError: Method Date.prototype.setMilliseconds called on incompatible receiver [object Object].
PASS o.__proto__.setSeconds() threw exception TypeError: Method Date.prototype.setSeconds called on incompatible receiver [object Object].
PASS o.__proto__.setMinutes() threw exception TypeError: Method Date.prototype.setMinutes called on incompatible receiver [object Object].
PASS o.__proto__.setHours() threw exception TypeError: Method Date.prototype.setHours called on incompatible receiver [object Object].
PASS o.__proto__.setDate() threw exception TypeError: Method Date.prototype.setDate called on incompatible receiver [object Object].
PASS o.__proto__.setMonth() threw exception TypeError: Method Date.prototype.setMonth called on incompatible receiver [object Object].
PASS o.__proto__.setFullYear() threw exception TypeError: Method Date.prototype.setFullYear called on incompatible receiver [object Object].
PASS o.__proto__.setYear() threw exception TypeError: Method Date.prototype.setYear called on incompatible receiver [object Object].
PASS successfullyParsed is true

TEST COMPLETE
